socket2

Macro impl_debug

macro_rules! impl_debug {
    (
        // Type name for which to implement `fmt::Debug`.
        $type: path,
        $(
            $(#[$target: meta])*
            // The flag(s) to check.
            // Need to specific the libc crate because Windows doesn't use
            // `libc` but `windows_sys`.
            $libc: ident :: $flag: ident
        ),+ $(,)*
    ) => { ... };
}

Macro to implement fmt::Debug for a type, printing the constant names rather than a number.

Note this is used in the sys module and thus must be defined before defining the modules.