Struct socket2::Protocol

pub struct Protocol(_);

Protocol specification used for creating sockets via Socket::new.

This is a newtype wrapper around an integer which provides a nicer API in addition to an injection point for documentation.

This type is freely interconvertible with C’s int type, however, if a raw value needs to be provided.

Implementations

impl Protocol

pub const ICMPV4: Protocol = _

Protocol corresponding to ICMPv4.

pub const ICMPV6: Protocol = _

Protocol corresponding to ICMPv6.

pub const TCP: Protocol = _

Protocol corresponding to TCP.

pub const UDP: Protocol = _

Protocol corresponding to UDP.