Struct socket2::Type

pub struct Type(_);

Specification of communication semantics on a socket.

This is a newtype wrapper around an integer which provides a nicer API in addition to an injection point for documentation. Convenience constructors such as Type::stream, Type::dgram, etc, are provided to avoid reaching into libc for various constants.

This type is freely interconvertible with the i32 type, however, if a raw value needs to be provided.

Methods

impl Type

pub fn stream() -> Type

Type corresponding to SOCK_STREAM

Used for protocols such as TCP.

pub fn dgram() -> Type

Type corresponding to SOCK_DGRAM

Used for protocols such as UDP.

pub fn seqpacket() -> Type

Type corresponding to SOCK_SEQPACKET

pub fn raw() -> Type

Type corresponding to SOCK_RAW