socket2

Module sys

Macros

syscall 🔒

Helper macro to execute a system call that returns an io::Result.

Structs

CcidEndpoints

See Socket::dccp_available_ccids.

IpMreq 🔒

IpMreqSource 🔒

Ipv6Mreq 🔒

MaybeUninitSlice

SockFilter

Berkeley Packet Filter (BPF).

linger 🔒

msghdr 🔒

sockaddr 🔒

sockaddr_in 🔒

sockaddr_in6 🔒

sockaddr_storage 🔒

Constants

AF_INET 🔒

AF_INET6 🔒

AF_UNIX 🔒

IPPROTO_DCCP 🔒

DCCP

IPPROTO_ICMP 🔒

IPPROTO_ICMPV6 🔒

IPPROTO_IP 🔒

IPPROTO_IPV6 🔒

IPPROTO_MPTCP 🔒

Multipath TCP

IPPROTO_SCTP 🔒

SCTP

IPPROTO_TCP 🔒

IPPROTO_UDP 🔒

IPPROTO_UDPLITE 🔒

IPV6_ADD_MEMBERSHIP 🔒

IPV6_DROP_MEMBERSHIP 🔒

IPV6_HDRINCL 🔒

IPV6_MULTICAST_HOPS 🔒

IPV6_MULTICAST_IF 🔒

IPV6_MULTICAST_LOOP 🔒

IPV6_RECVHOPLIMIT 🔒

IPV6_RECVTCLASS 🔒

IPV6_UNICAST_HOPS 🔒

IPV6_V6ONLY 🔒

IP_ADD_MEMBERSHIP 🔒

IP_ADD_SOURCE_MEMBERSHIP 🔒

IP_DROP_MEMBERSHIP 🔒

IP_DROP_SOURCE_MEMBERSHIP 🔒

IP_HDRINCL 🔒

IP_MULTICAST_IF 🔒

IP_MULTICAST_LOOP 🔒

IP_MULTICAST_TTL 🔒

IP_RECVTOS 🔒

IP_TOS 🔒

IP_TTL 🔒

MAX_BUF_LEN 🔒

Maximum size of a buffer passed to system call like recv and send.

MSG_OOB 🔒

MSG_PEEK 🔒

MSG_TRUNC 🔒

SOCK_DCCP 🔒

SOCK_DGRAM 🔒

SOCK_RAW 🔒

SOCK_SEQPACKET 🔒

SOCK_STREAM 🔒

SOL_SOCKET 🔒

SO_BROADCAST 🔒

SO_ERROR 🔒

SO_KEEPALIVE 🔒

SO_LINGER 🔒

SO_OOBINLINE 🔒

SO_PASSCRED 🔒

SO_PRIORITY 🔒

SO_RCVBUF 🔒

SO_RCVTIMEO 🔒

SO_REUSEADDR 🔒

SO_SNDBUF 🔒

SO_SNDTIMEO 🔒

SO_TYPE 🔒

TCP_CA_NAME_MAX 🔒

TCP_KEEPCNT 🔒

TCP_KEEPINTVL 🔒

TCP_NODELAY 🔒

Functions

accept 🔒

bind 🔒

connect 🔒

fcntl_add 🔒

Add flag to the current set flags of F_GETFD.

fcntl_get 🔒

Get the flags using cmd.

fcntl_remove 🔒

Remove flag to the current set flags of F_GETFD.

from_in6_addr 🔒

from_in_addr 🔒

from_timeval 🔒

getpeername 🔒

getsockname 🔒

getsockopt 🔒 ⚠

Caller must ensure T is the correct type for opt and val.

into_secs 🔒

into_timeval 🔒

listen 🔒

msghdr_control_len 🔒

msghdr_flags 🔒

nonblocking 🔒

offset_of_path 🔒

Returns the offset of the sun_path member of the passed unix socket address.

original_dst_v4 🔒

original_dst_v6 🔒

Get the value for the IP6T_SO_ORIGINAL_DST option on this socket.

peek_sender 🔒

poll_connect 🔒

recv 🔒

recv_from 🔒

recv_from_vectored 🔒

recv_vectored 🔒

recvmsg 🔒

send 🔒

send_to 🔒

send_to_vectored 🔒

send_vectored 🔒

sendmsg 🔒

set_msghdr_control 🔒

set_msghdr_flags 🔒

set_msghdr_iov 🔒

set_msghdr_name 🔒

set_nonblocking 🔒

set_tcp_keepalive 🔒

set_timeout_opt 🔒

Wrapper around setsockopt to deal with platform specific timeouts.

setsockopt 🔒 ⚠

Caller must ensure T is the correct type for opt and val.

shutdown 🔒

socket 🔒

socket_as_raw 🔒

socket_from_raw 🔒 ⚠

socket_into_raw 🔒

socketpair 🔒

tcp_keepalive_time 🔒

timeout_opt 🔒

Wrapper around getsockopt to deal with platform specific timeouts.

to_in6_addr 🔒

to_in_addr 🔒

to_mreqn 🔒

try_clone 🔒

unix_sockaddr 🔒

Type Aliases

Bool 🔒

IovLen 🔒

RawSocket 🔒

Socket 🔒

c_int 🔒

Equivalent to C’s signed int (int) type.

sa_family_t 🔒

socklen_t 🔒