Struct socket2::RecvFlags

pub struct RecvFlags(_);

Flags for incoming messages.

Flags provide additional information about incoming messages.

Implementations

impl RecvFlags

Unix-only API.

pub const fn is_end_of_record(self) -> bool

Check if the message terminates a record.

Not all socket types support the notion of records. For socket types that do support it (such as SEQPACKET), a record is terminated by sending a message with the end-of-record flag set.

On Unix this corresponds to the MSG_EOR flag.

pub const fn is_out_of_band(self) -> bool

Check if the message contains out-of-band data.

This is useful for protocols where you receive out-of-band data mixed in with the normal data stream.

On Unix this corresponds to the MSG_OOB flag.

impl RecvFlags

pub const fn is_truncated(self) -> bool

Check if the message contains a truncated datagram.

This flag is only used for datagram-based sockets, not for stream sockets.

On Unix this corresponds to the MSG_TRUNC flag. On Windows this corresponds to the WSAEMSGSIZE error code.