Struct socket2::SockRef

pub struct SockRef<'s> { /* fields omitted */ }

A reference to a Socket that can be used to configure socket types other than the Socket type itself.

This allows for example a TcpStream, found in the standard library, to be configured using all the additional methods found in the Socket API.

SockRef can be created from any socket type that implements AsRawFd (Unix) or AsRawSocket (Windows) using the From implementation, but the caller must ensure the file descriptor/socket is a valid.

Examples

use std::net::{TcpStream, SocketAddr};

use socket2::SockRef;

// Create `TcpStream` from the standard library.
let address: SocketAddr = "127.0.0.1:1234".parse()?;
let stream = TcpStream::connect(address)?;

// Create a `SockRef`erence to the stream.
let socket_ref = SockRef::from(&stream);
// Use `Socket::set_nodelay` on the stream.
socket_ref.set_nodelay(true)?;
drop(socket_ref);

assert_eq!(stream.nodelay()?, true);