Module socket2::sys

Macros

syscall ğŸ”’

Helper macro to execute a system call that returns an io::Result.

Structs

CcidEndpoints

See [Socket::dccp_available_ccids].

IpMreq ğŸ”’

IpMreqSource ğŸ”’

Ipv6Mreq ğŸ”’

MaybeUninitSlice

linger ğŸ”’

msghdr ğŸ”’

sockaddr ğŸ”’

sockaddr_in ğŸ”’

sockaddr_in6 ğŸ”’

sockaddr_storage ğŸ”’

Constants

AF_INET ğŸ”’

AF_INET6 ğŸ”’

AF_UNIX ğŸ”’

IPPROTO_DCCP ğŸ”’

DCCP

IPPROTO_ICMP ğŸ”’

IPPROTO_ICMPV6 ğŸ”’

IPPROTO_IP ğŸ”’

IPPROTO_IPV6 ğŸ”’

IPPROTO_MPTCP ğŸ”’

Multipath TCP

IPPROTO_SCTP ğŸ”’

SCTP

IPPROTO_TCP ğŸ”’

IPPROTO_UDP ğŸ”’

IPPROTO_UDPLITE ğŸ”’

IPV6_ADD_MEMBERSHIP ğŸ”’

IPV6_DROP_MEMBERSHIP ğŸ”’

IPV6_MULTICAST_HOPS ğŸ”’

IPV6_MULTICAST_IF ğŸ”’

IPV6_MULTICAST_LOOP ğŸ”’

IPV6_RECVTCLASS ğŸ”’

IPV6_UNICAST_HOPS ğŸ”’

IPV6_V6ONLY ğŸ”’

IP_ADD_MEMBERSHIP ğŸ”’

IP_ADD_SOURCE_MEMBERSHIP ğŸ”’

IP_DROP_MEMBERSHIP ğŸ”’

IP_DROP_SOURCE_MEMBERSHIP ğŸ”’

IP_HDRINCL ğŸ”’

IP_MULTICAST_IF ğŸ”’

IP_MULTICAST_LOOP ğŸ”’

IP_MULTICAST_TTL ğŸ”’

IP_RECVTOS ğŸ”’

IP_TOS ğŸ”’

IP_TTL ğŸ”’

MAX_BUF_LEN ğŸ”’

Maximum size of a buffer passed to system call like recv and send.

MSG_OOB ğŸ”’

MSG_PEEK ğŸ”’

MSG_TRUNC ğŸ”’

SOCK_DCCP ğŸ”’

SOCK_DGRAM ğŸ”’

SOCK_RAW ğŸ”’

SOCK_SEQPACKET ğŸ”’

SOCK_STREAM ğŸ”’

SOL_SOCKET ğŸ”’

SO_BROADCAST ğŸ”’

SO_ERROR ğŸ”’

SO_KEEPALIVE ğŸ”’

SO_LINGER ğŸ”’

SO_OOBINLINE ğŸ”’

SO_RCVBUF ğŸ”’

SO_RCVTIMEO ğŸ”’

SO_REUSEADDR ğŸ”’

SO_SNDBUF ğŸ”’

SO_SNDTIMEO ğŸ”’

SO_TYPE ğŸ”’

TCP_CA_NAME_MAX ğŸ”’

TCP_KEEPCNT ğŸ”’

TCP_KEEPINTVL ğŸ”’

TCP_NODELAY ğŸ”’

Functions

accept ğŸ”’

bind ğŸ”’

connect ğŸ”’

fcntl_add ğŸ”’

Add flag to the current set flags of F_GETFD.

fcntl_get ğŸ”’

Get the flags using cmd.

fcntl_remove ğŸ”’

Remove flag to the current set flags of F_GETFD.

from_in6_addr ğŸ”’

from_in_addr ğŸ”’

from_timeval ğŸ”’

getpeername ğŸ”’

getsockname ğŸ”’

getsockopt ğŸ”’ âš

Caller must ensure T is the correct type for opt and val.

into_secs ğŸ”’

into_timeval ğŸ”’

keepalive_time ğŸ”’

listen ğŸ”’

msghdr_flags ğŸ”’

nonblocking ğŸ”’

offset_of_path ğŸ”’

Returns the offset of the sun_path member of the passed unix socket address.

peek_sender ğŸ”’

poll_connect ğŸ”’

recv ğŸ”’

recv_from ğŸ”’

recv_from_vectored ğŸ”’

recv_vectored ğŸ”’

recvmsg ğŸ”’

send ğŸ”’

send_to ğŸ”’

send_to_vectored ğŸ”’

send_vectored ğŸ”’

sendmsg ğŸ”’

set_msghdr_control ğŸ”’

set_msghdr_flags ğŸ”’

set_msghdr_iov ğŸ”’

set_msghdr_name ğŸ”’

set_nonblocking ğŸ”’

set_tcp_keepalive ğŸ”’

set_timeout_opt ğŸ”’

Wrapper around setsockopt to deal with platform specific timeouts.

setsockopt ğŸ”’ âš

Caller must ensure T is the correct type for opt and val.

shutdown ğŸ”’

socket ğŸ”’

socket_as_raw ğŸ”’

socket_from_raw ğŸ”’ âš

socket_into_raw ğŸ”’

socketpair ğŸ”’

timeout_opt ğŸ”’

Wrapper around getsockopt to deal with platform specific timeouts.

to_in6_addr ğŸ”’

to_in_addr ğŸ”’

to_mreqn ğŸ”’

try_clone ğŸ”’

unix_sockaddr ğŸ”’

Type Aliases

Bool ğŸ”’

IovLen ğŸ”’

Socket ğŸ”’

c_int ğŸ”’

sa_family_t ğŸ”’

socklen_t ğŸ”’