Module socket2::sys

Macros

syscall

Helper macro to execute a system call that returns an io::Result.

Structs

IpMreq

Ipv6Mreq

MaybeUninitSlice

linger

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_storage

Constants

AF_INET

AF_INET6

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_TCP

IPPROTO_UDP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_LOOP

IPV6_UNICAST_HOPS

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TOS

IP_TTL

MAX_BUF_LEN

Maximum size of a buffer passed to system call like recv and send.

MSG_OOB

MSG_PEEK

MSG_TRUNC

SOCK_DGRAM

SOCK_STREAM

SOL_SOCKET

SO_BROADCAST

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_OOBINLINE

SO_RCVBUF

SO_RCVTIMEO

SO_REUSEADDR

SO_SNDBUF

SO_SNDTIMEO

SO_TYPE

TCP_NODELAY

Functions

accept

bind

clamp

connect

fcntl_add

Add flag to the current set flags of F_GETFD.

fcntl_remove

Remove flag to the current set flags of F_GETFD.

from_in6_addr

from_in_addr

from_timeval

getpeername

getsockname

getsockopt⚠

Caller must ensure T is the correct type for opt and val.

into_secs

into_timeval

listen

poll_connect

recv

recv_from

recv_from_vectored

recv_vectored

recvmsg

Returns the (bytes received, sending address len, RecvFlags).

send

send_to

send_to_vectored

send_vectored

sendmsg

Returns the (bytes received, sending address len, RecvFlags).

set_nonblocking

set_tcp_keepalive

set_timeout_opt

Wrapper around setsockopt to deal with platform specific timeouts.

setsockopt⚠

Caller must ensure T is the correct type for opt and val.

shutdown

socket

socket_as_raw

socket_from_raw⚠

socket_into_raw

timeout_opt

Wrapper around getsockopt to deal with platform specific timeouts.

to_in6_addr

to_in_addr

try_clone

Type Definitions

Bool

IovLen

Socket

c_int

sa_family_t

socklen_t